+++ itou/www/apply/urls.py
@@ -1,7 +1,11 @@
 from django.urls import path
 
 from itou.www.apply.views import edit_views, list_views, process_views, submit_views
-from itou.www.job_seekers_views.views import CheckNIRForJobSeekerView, CheckNIRForSenderView
+from itou.www.job_seekers_views.views import (
+    CheckNIRForJobSeekerView,
+    CheckNIRForSenderView,
+    SearchByEmailForSenderView,
+)
 
 
 # https://docs.djangoproject.com/en/dev/topics/http/urls/#url-namespaces-and-included-urlconfs
@@ -18,9 +22,10 @@ urlpatterns = [
     ),
     # Ewen: deprecated url. These urls are going to job_seekers_views.views
     path("<int:company_pk>/sender/check_nir", CheckNIRForSenderView.as_view(), name="check_nir_for_sender"),
+    # Ewen: deprecated url. These urls are going to job_seekers_views
     path(
         "<int:company_pk>/sender/search-by-email/<uuid:session_uuid>",
-        submit_views.SearchByEmailForSenderView.as_view(),
+        SearchByEmailForSenderView.as_view(),
         name="search_by_email_for_sender",
     ),
     path(
@@ -115,9 +120,10 @@ urlpatterns = [
         name="check_nir_for_hire",
         kwargs={"hire_process": True},
     ),
+    # Ewen: deprecated url. These urls are going to job_seekers_views
     path(
         "<int:company_pk>/hire/search-by-email/<uuid:session_uuid>",
-        submit_views.SearchByEmailForSenderView.as_view(),
+        SearchByEmailForSenderView.as_view(),
         name="search_by_email_for_hire",
         kwargs={"hire_process": True},
     ),
