+++ itou/www/apply/views/submit_views.py
@@ -13,7 +13,6 @@ from django.http import Http404, HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
 from django.utils import timezone
-from django.utils.html import format_html
 from django.views.generic import TemplateView
 
 from itou.approvals.models import Approval
@@ -37,7 +36,6 @@ from itou.www.apply.forms import (
     CreateOrUpdateJobSeekerStep1Form,
     CreateOrUpdateJobSeekerStep2Form,
     CreateOrUpdateJobSeekerStep3Form,
-    JobSeekerExistsForm,
     SubmitJobApplicationForm,
 )
 from itou.www.apply.views import common as common_views, constants as apply_view_constants
@@ -310,108 +308,6 @@ class PendingAuthorizationForSender(ApplyStepForSenderBaseView):
     template_name = "apply/submit_step_pending_authorization.html"
 
 
-class SearchByEmailForSenderView(SessionNamespaceRequiredMixin, ApplyStepForSenderBaseView):
-    required_session_namespaces = ["job_seeker_session"]
-    template_name = "job_seekers_views/step_search_job_seeker_by_email.html"
-
-    def __init__(self):
-        super().__init__()
-        self.form = None
-
-    def setup(self, request, *args, **kwargs):
-        self.job_seeker_session = SessionNamespace(request.session, kwargs["session_uuid"])
-        super().setup(request, *args, **kwargs)
-        self.form = JobSeekerExistsForm(
-            is_gps=self.is_gps, initial=self.job_seeker_session.get("user", {}), data=request.POST or None
-        )
-
-    def post(self, request, *args, **kwargs):
-        can_add_nir = False
-        preview_mode = False
-        job_seeker = None
-
-        if self.form.is_valid():
-            job_seeker = self.form.get_user()
-            nir = self.job_seeker_session.get("profile", {}).get("nir")
-            can_add_nir = nir and self.sender.can_add_nir(job_seeker)
-
-            # No user found with that email, redirect to create a new account.
-            if not job_seeker:
-                user_infos = self.job_seeker_session.get("user", {})
-                user_infos.update({"email": self.form.cleaned_data["email"]})
-                profile_infos = self.job_seeker_session.get("profile", {})
-                profile_infos.update({"nir": nir})
-                self.job_seeker_session.update({"user": user_infos, "profile": profile_infos})
-                view_name = (
-                    "apply:create_job_seeker_step_1_for_hire"
-                    if self.hire_process
-                    else "apply:create_job_seeker_step_1_for_sender"
-                )
-
-                return HttpResponseRedirect(
-                    reverse(
-                        view_name, kwargs={"company_pk": self.company.pk, "session_uuid": self.job_seeker_session.name}
-                    )
-                    + ("?gps=true" if self.is_gps else "")
-                )
-
-            # Ask the sender to confirm the email we found is associated to the correct user
-            if self.form.data.get("preview"):
-                preview_mode = True
-
-            # The email we found is correct
-            if self.form.data.get("confirm"):
-                if not can_add_nir:
-                    return self.redirect_to_check_infos(job_seeker.public_id)
-
-                try:
-                    job_seeker.jobseeker_profile.nir = nir
-                    job_seeker.jobseeker_profile.lack_of_nir_reason = ""
-                    job_seeker.jobseeker_profile.save(update_fields=["nir", "lack_of_nir_reason"])
-                except ValidationError:
-                    msg = format_html(
-                        "Le<b> numéro de sécurité sociale</b> renseigné ({}) est "
-                        "déjà utilisé par un autre candidat sur la Plateforme.<br>"
-                        "Merci de renseigner <b>le numéro personnel et unique</b> "
-                        "du candidat pour lequel vous souhaitez postuler.",
-                        nir,
-                    )
-                    messages.warning(request, msg)
-                    logger.exception("step_job_seeker: error when saving job_seeker=%s nir=%s", job_seeker, nir)
-                else:
-                    if self.is_gps:
-                        FollowUpGroup.objects.follow_beneficiary(
-                            beneficiary=job_seeker, user=request.user, is_referent=True
-                        )
-                        return HttpResponseRedirect(reverse("gps:my_groups"))
-                    else:
-                        return self.redirect_to_check_infos(job_seeker.public_id)
-
-        return self.render_to_response(
-            self.get_context_data(**kwargs)
-            | {
-                "can_add_nir": can_add_nir,
-                "preview_mode": preview_mode,
-                "job_seeker": job_seeker,
-                "can_view_personal_information": job_seeker and self.sender.can_view_personal_information(job_seeker),
-            }
-        )
-
-    def get_back_url(self):
-        view_name = (
-            "job_seekers_views:check_nir_for_hire" if self.hire_process else "job_seekers_views:check_nir_for_sender"
-        )
-        return reverse(view_name, kwargs={"company_pk": self.company.pk})
-
-    def get_context_data(self, **kwargs):
-        return super().get_context_data(**kwargs) | {
-            "form": self.form,
-            "nir": self.job_seeker_session.get("profile", {}).get("nir"),
-            "siae": self.company,
-            "preview_mode": False,
-        }
-
-
 class CreateJobSeekerForSenderBaseView(SessionNamespaceRequiredMixin, ApplyStepForSenderBaseView):
     required_session_namespaces = ["job_seeker_session"]
 
@@ -441,8 +337,8 @@ class CreateJobSeekerForSenderBaseView(SessionNamespaceRequiredMixin, ApplyStepF
 class CreateJobSeekerStep1ForSenderView(CreateJobSeekerForSenderBaseView):
     template_name = "job_seekers_views/create_or_update_job_seeker/step_1.html"
 
-    previous_apply_url = "apply:search_by_email_for_sender"
-    previous_hire_url = "apply:search_by_email_for_hire"
+    previous_apply_url = "job_seekers_views:search_by_email_for_sender"
+    previous_hire_url = "job_seekers_views:search_by_email_for_hire"
     next_apply_url = "apply:create_job_seeker_step_2_for_sender"
     next_hire_url = "apply:create_job_seeker_step_2_for_hire"
